+++ main.py
@@ -11,27 +11,31 @@ load_dotenv()
 TOKEN = os.getenv('DISCORD_TOKEN')
 
 bot = commands.Bot(command_prefix='!')
+bot.remove_command('help')
 
 @bot.event
 async def on_ready():
   print(f'{bot.user.name} has connected to Discord!')
   await bot.change_presence(activity=discord.Game('skribbl.io'))
 
-@bot.command(name="play", help='send link to Skribbl.io')
-async def play(ctx):
-  await ctx.send("https://skribbl.io/")
+# @bot.command(name="play", help='send link to Skribbl.io')
+# async def play(ctx):
+#   await ctx.send("https://skribbl.io/")
 
-@bot.command(name="all-words", help="return file with all the words in the DB")
-async def get_all_words(ctx):
-  await ctx.send('making the file')
-  words = helper.get_all_words()
+async def send_words_file(ctx, words):
   if words:
     with open("words.txt", "w") as file:
-      file.write(', '.join(words))
+      file.write(','.join(words))
     with open("words.txt", "rb") as file:
       await ctx.send("Your file is:", file=discord.File(file, "words.txt"))
   else:
     await ctx.send('no words has been found')
+
+@bot.command(name="all", help="return file with all the words in the DB")
+async def get_all_words(ctx):
+  await ctx.send('making the file')
+  words = helper.get_all_words()
+  await send_words_file(ctx, words)
 
 @bot.command(name="all-indexed", help="return indexed file of all the words")
 async def get_all_words_sorted(ctx):
@@ -45,41 +49,23 @@ async def get_all_words_sorted(ctx):
   else:
     await ctx.send('no words has been found')
 
-@bot.command(name="first-words", help='argu: number:int return file with the {number} first words')
+@bot.command(name="first", help='argu: number:int return file with the {number} first words')
 async def get_first_words(ctx, number:int):
   await ctx.send('making the file')
   words = helper.get_first_words(number)
-  if words:
-    with open("words.txt", "w") as file:
-      file.write(', '.join(words))
-    with open("words.txt", "rb") as file:
-      await ctx.send("Your file is:", file=discord.File(file, "words.txt"))
-  else:
-    await ctx.send('no words has been found')
+  await send_words_file(ctx, words)
 
-@bot.command(name="last-words", help='argu: number:int return file with the {number} last words')
+@bot.command(name="last", help='argu: number:int return file with the {number} last words')
 async def get_last_words(ctx, number:int):
   await ctx.send('making the file')
   words = helper.get_last_words(number)
-  if words:
-    with open("words.txt", "w") as file:
-      file.write(', '.join(words))
-    with open("words.txt", "rb") as file:
-      await ctx.send("Your file is:", file=discord.File(file, "words.txt"))
-  else:
-    await ctx.send('no words has been found')
+  await send_words_file(ctx, words)
 
-@bot.command(name="random-words", help='argu: number:int return file with {number} random words')
-async def get_random_words(ctx, number:int):
+@bot.command(name="random", help='argu: number:int last(optional); return file with {number} random words and {last} last words')
+async def get_random_words(ctx, number:int, last = 0):
   await ctx.send('making the file')
-  words = helper.get_random_words(number)
-  if words:
-    with open("words.txt", "w") as file:
-      file.write(', '.join(words))
-    with open("words.txt", "rb") as file:
-      await ctx.send("Your file is:", file=discord.File(file, "words.txt"))
-  else:
-    await ctx.send('no words has been found')
+  words = helper.get_random_words(number, last)
+  await send_words_file(ctx, words)
 
 @bot.command(name="words-in-chat", help="write 75 words in chat")
 async def get_words_in_chat(ctx):
@@ -90,29 +76,23 @@ async def get_words_in_chat(ctx):
   else:
     await ctx.send('no words has been found')
 
-@bot.command(name="shuffled-last-words", help='argu: last:int, random:int return file with {last} last words and {random} random words')
-async def get_last_and_random_words(ctx, last_amount:int, random_amount:int):
+@bot.command(name="first-last", help="return file with {number} first and last words")
+async def get_first_last(ctx, number:int):
   await ctx.send('making the file')
-  words = helper.get_last_and_random_words(last_amount, random_amount)
-  if words:
-    with open("words.txt", "w") as file:
-      file.write(', '.join(words))
-    with open("words.txt", "rb") as file:
-      await ctx.send("Your file is:", file=discord.File(file, "words.txt"))
-  else:
-    await ctx.send('no words has been found')
+  words = helper.get_first_last(number)
+  await send_words_file(ctx, words)
 
-@bot.command(name="words-amount", help='print the amount of words in the DB')
+@bot.command(name="count-all", help='print the amount of words in the DB')
 async def get_words_number(ctx):
   num = helper.count_words()
   await ctx.send(f'we have **{num}** words!')
 
-@bot.command(name="count-new-words", help='print the amount of words that added in the past 48 hours')
+@bot.command(name="count-new", help='print the amount of words that added in the past 48 hours')
 async def count_new_words(ctx):
   num = helper.count_new_words()
   await ctx.send(f'**{num}** words added')
 
-@bot.command(name='add-word', help='argu: arg:str add the coma separeted words in {arg} to the DB')
+@bot.command(name='add', help='argu: arg:str add the coma separeted words in {arg} to the DB')
 async def add_words(ctx, *, arg):
   MAX_WORD_LENGTH = 30
   word_list = arg.split(',')
@@ -122,7 +102,7 @@ async def add_words(ctx, *, arg):
   for word in word_list:
     word = word.replace('\'', '')
     if(len(word) > MAX_WORD_LENGTH):
-      await ctx.send(f'{word} is to long, maximum characters allowed is ' + str(MAX_WORD_LENGTH) + " chracters")
+      await ctx.send(f'{word} it is to long, maximum characters allowed is ' + str(MAX_WORD_LENGTH) + " chracters")
     elif bool(re.match('[א-ת0-9\s?!]+$', word)):
       helper.__add_word__(word, author.name, msg_date)
       count_words += 1
@@ -135,13 +115,43 @@ async def add_words(ctx, *, arg):
   except:
     pass
 
-@bot.command(name='word-author', help='argu: word:str print the word author')
+@bot.command(name='author', help='argu: word:str print the word author')
 async def get_author(ctx, *, word):
   author = helper.get_author(word)
   if author is not None:
     await ctx.send(f'**{author}** is the author')
   else:
-    await ctx.send("Word is not existing")
+    await ctx.send("Word not found")
+
+@bot.command(name="help", help="show this message")
+async def help(ctx, command=None):
+  if(command is None):
+    embed = discord.Embed(
+      title="help", colour= discord.Colour.blue()
+    )
+    embed.set_footer(text="Type !help command for more info on a command.")
+
+    for com in bot.commands:
+      # print(com.signature)
+      embed.add_field(name=com.name, value=com.help, inline=False)
+
+    await ctx.send(embed = embed)
+  else:
+    com = bot.get_command(command)
+    if com is not None:
+      embed = discord.Embed(
+        title=com.name, colour= discord.Colour.blue()
+      )
+      params = ""
+      for v in com.clean_params.items():
+       params += str(v[1]) + ", "
+      if params == "":
+        params = "no parameters"
+      embed.add_field(name=params, value=com.help)
+
+      await ctx.send(embed = embed)
+    else:
+      await ctx.send("command not found")
 
 @bot.event
 async def on_command_error(ctx, error):
@@ -150,10 +160,10 @@ async def on_command_error(ctx, error):
   elif isinstance(error, commands.errors.MissingRequiredArgument):
     await ctx.send(f'missing command argument: {error.param}')
   elif isinstance(error, commands.errors.BadArgument):
-    await ctx.send(f'parameter must be number')
+    await ctx.send('parameter must be number')
   elif isinstance(error, commands.errors.CommandNotFound):
     print('command not found')
-    await ctx.send(f'command not found')
+    await ctx.send('command not found')
   else:
     await ctx.send('some error occured with the command')
     guild = ctx.guild